Move getNextFocusableElement into its own file
