+++ src/TreeView/getNextFocusableElement.ts
@@ -1,0 +1,135 @@
+// DOM utilities used for focus management
+
+export function getNextFocusableElement(
+  activeElement: HTMLElement,
+  event: React.KeyboardEvent<HTMLElement>
+): HTMLElement | undefined {
+  const elementState = getElementState(activeElement)
+
+  // Reference: https://www.w3.org/WAI/ARIA/apg/patterns/treeview/#keyboard-interaction-24
+  switch (`${elementState} ${event.key}`) {
+    case 'open ArrowRight':
+      // Focus first child node
+      return getFirstChildElement(activeElement)
+
+    case 'open ArrowLeft':
+      // Close node; don't change focus
+      return
+
+    case 'closed ArrowRight':
+      // Open node; don't change focus
+      return
+
+    case 'closed ArrowLeft':
+      // Focus parent element
+      return getParentElement(activeElement)
+
+    case 'end ArrowRight':
+      // Do nothing
+      return
+
+    case 'end ArrowLeft':
+      // Focus parent element
+      return getParentElement(activeElement)
+  }
+
+  // ArrowUp, ArrowDown, Home, and End behavior are the same regarless of element state
+  switch (event.key) {
+    case 'ArrowUp':
+      // Focus previous visible element
+      return getVisibleElement(activeElement, 'previous')
+
+    case 'ArrowDown':
+      // Focus next visible element
+      return getVisibleElement(activeElement, 'next')
+
+    case 'Home':
+      // Focus first visible element
+      return getFirstElement(activeElement)
+
+    case 'End':
+      // Focus last visible element
+      return getLastElement(activeElement)
+  }
+}
+
+function getElementState(element: HTMLElement): 'open' | 'closed' | 'end' {
+  if (element.getAttribute('role') !== 'treeitem') {
+    throw new Error('Element is not a treeitem')
+  }
+
+  switch (element.getAttribute('aria-expanded')) {
+    case 'true':
+      return 'open'
+    case 'false':
+      return 'closed'
+    default:
+      return 'end'
+  }
+}
+
+function getVisibleElement(element: HTMLElement, direction: 'next' | 'previous'): HTMLElement | undefined {
+  const root = element.closest('[role=tree]')
+
+  if (!root) return
+
+  const walker = document.createTreeWalker(root, NodeFilter.SHOW_ELEMENT, node => {
+    if (!(node instanceof HTMLElement)) return NodeFilter.FILTER_SKIP
+    return node.getAttribute('role') === 'treeitem' ? NodeFilter.FILTER_ACCEPT : NodeFilter.FILTER_SKIP
+  })
+
+  let current = walker.firstChild()
+
+  while (current !== element) {
+    current = walker.nextNode()
+  }
+
+  let next = direction === 'next' ? walker.nextNode() : walker.previousNode()
+
+  // If next element is nested inside a collapsed subtree, continue iterating
+  while (next instanceof HTMLElement && next.parentElement?.closest('[role=treeitem][aria-expanded=false]')) {
+    next = direction === 'next' ? walker.nextNode() : walker.previousNode()
+  }
+
+  return next instanceof HTMLElement ? next : undefined
+}
+
+function getFirstChildElement(element: HTMLElement): HTMLElement | undefined {
+  const firstChild = element.querySelector('[role=treeitem]')
+  return firstChild instanceof HTMLElement ? firstChild : undefined
+}
+
+function getParentElement(element: HTMLElement): HTMLElement | undefined {
+  const group = element.closest('[role=group]')
+  const parent = group?.closest('[role=treeitem]')
+  return parent instanceof HTMLElement ? parent : undefined
+}
+
+function getFirstElement(element: HTMLElement): HTMLElement | undefined {
+  const root = element.closest('[role=tree]')
+  const first = root?.querySelector('[role=treeitem]')
+  return first instanceof HTMLElement ? first : undefined
+}
+
+function getLastElement(element: HTMLElement): HTMLElement | undefined {
+  const root = element.closest('[role=tree]')
+  const items = Array.from(root?.querySelectorAll('[role=treeitem]') || [])
+
+  // If there are no items, return undefined
+  if (items.length === 0) return
+
+  let index = items.length - 1
+  let last = items[index]
+
+  // If last element is nested inside a collapsed subtree, continue iterating
+  while (
+    index > 0 &&
+    last instanceof HTMLElement &&
+    last.parentElement?.closest('[role=treeitem][aria-expanded=false]')
+  ) {
+    index -= 1
+    last = items[index]
+  }
+
+  return last instanceof HTMLElement ? last : undefined
+}
